+++ wp-content/themes/footmate/app/Integrations/Integrations.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace FM\Integrations;
+
+use FM\Integrations\ESPN;
+
+class Integrations
+{
+    private ESPN $espn;
+
+    public function __construct()
+    {
+        $this->espn = new ESPN();
+    }
+
+    public function espn(): ESPN
+    {
+        return $this->espn;
+    }
+}