+++ wp-content/themes/footmate/functions.php
@@ -1,9 +1,9 @@
 <?php
 
-define('FM_VERSION', '0.1.0');
+define('FM_VERSION', '0.1.1');
 define('FM_PATH', dirname(__FILE__));
 define('FM_FILE', FM_PATH . '/functions.php');
 
 require_once(FM_PATH . '/inc/bootstrap.php');
 
-new FM\App();
+fm();